lib/shadow/, lib/, src/: struct sgrp: Move to separate file
