+++ src/userdel.c
@@ -50,6 +50,7 @@
 #ifdef ENABLE_SUBIDS
 #include "subordinateio.h"
 #endif				/* ENABLE_SUBIDS */
+#include "shadow/gshadow/sgrp.h"
 #include "shadowlog.h"
 #include "string/sprintf/aprintf.h"
 #include "string/sprintf/xaprintf.h"
